requirements_enhanced.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.0.0
tifffile>=2021.7.2
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0
opencv-python>=4.5.0
timm>=0.9.0